[CHORE] 독서 기록 레포지토리 분리
