+++ src/main/java/umc/nook/bookshelves/repository/BookshelfCustomRepositoryImpl.java
@@ -235,44 +235,6 @@ class BookShelfCustomRepositoryImpl implements BookShelfCustomRepository {
         return new BookShelfDTO.BooksInsightDTO(totalCount, recordCount, typeDTOs);
     }
 
-    @Transactional
-    public RecordDTO.MonthlyRecordRateResponseDTO viewRecordRate(User user, Year year) {
-        QUserBookShelf ub = QUserBookShelf.userBookShelf;
-
-        Map<Integer, List<UserBookShelf>> booksByMonth = queryFactory
-                .selectFrom(ub)
-                .where(
-                        ub.user.eq(user),
-                        ub.createdDate.year().eq(year.getValue())
-                )
-                .fetch()
-                .stream()
-                .collect(Collectors.groupingBy(b -> b.getCreatedDate().getMonthValue()));
-
-        List<RecordDTO.MonthlyRecordRateResponseDTO.MonthRate> rates = new ArrayList<>();
-
-        for (int month = 1; month <= 12; month++) {
-            List<UserBookShelf> booksInMonth = booksByMonth.getOrDefault(month, List.of());
-
-            if (booksInMonth.isEmpty()) {
-                rates.add(new RecordDTO.MonthlyRecordRateResponseDTO.MonthRate(month, 0.0));
-                continue;
-            }
-
-            long total = booksInMonth.size();
-            long recorded = booksInMonth.stream()
-                    .filter(b -> b.getRecordedAt() != null)
-                    .count();
-
-            double rate = (recorded * 100.0) / total;
-            rates.add(new RecordDTO.MonthlyRecordRateResponseDTO.MonthRate(month, rate));
-        }
-
-        return new RecordDTO.MonthlyRecordRateResponseDTO(rates);
-    }
-
-
-
