+++ src/main/java/umc/nook/records/repository/BookRecordRepository.java
@@ -9,7 +9,7 @@ import umc.nook.records.domain.BookRecord;
 import java.util.List;
 import java.util.Optional;
 
-public interface BookRecordRepository extends JpaRepository<BookRecord,Long> {
+public interface BookRecordRepository extends JpaRepository<BookRecord,Long> , RecordCustomRepository{
     List<BookRecord> findAllByBookshelfAndParentIsNullOrderByCreatedDateAsc(UserBookShelf userBookShelf);
 
     List<BookRecord> findAllByParent(BookRecord record);
