+++ src/main/java/umc/nook/records/repository/RecordCustomRepository.java
@@ -1,5 +1,11 @@
 package umc.nook.records.repository;
 
+import umc.nook.records.dto.RecordDTO;
+import umc.nook.users.domain.User;
+
+import java.time.Year;
+
 public interface RecordCustomRepository {
+    RecordDTO.MonthlyRecordRateResponseDTO viewRecordRate(User user, Year year);
 
 }
